fix(pages): 리뷰 이전 refactor
