+++ packages/admin/src/App.tsx
@@ -6,7 +6,6 @@ export default function App() {
 	return (
 		<AppProviders>
 			{/* TODO: create global loading component */}
-
 			<RouterProvider router={router} fallbackElement={<>loading...</>} />
 		</AppProviders>
 	);
